Move login status to new file
